requirements.txt
llama-index
weaviate
sentence-transformer
transformers
python-telegram-bot
